group signature logic in workflow
