+++ src/main/java/es/in2/issuer/application/workflow/CredentialSignerWorkflow.java
@@ -3,5 +3,6 @@ package es.in2.issuer.application.workflow;
 import reactor.core.publisher.Mono;
 
 public interface CredentialSignerWorkflow {
-    Mono<Void> signCredential(String authorizationHeader, String procedureId);
+    Mono<String> signAndUpdateCredential(String authorizationHeader, String procedureId);
+    Mono<String> signCredentialOnRequestedFormat(String unsignedCredential, String format, String token);
 }
