+++ src/main/java/es/in2/issuer/application/workflow/VerifiableCredentialIssuanceWorkflow.java
@@ -14,11 +14,5 @@ public interface VerifiableCredentialIssuanceWorkflow {
 
     Mono<VerifiableCredentialResponse> generateVerifiableCredentialDeferredResponse(String processId, DeferredCredentialRequest deferredCredentialRequest);
 
-    // Method for signing deferred credential using remote DSS, currently not in use in DOME profile
-    Mono<Void> signDeferredCredential(String unsignedCredential, String userId, UUID credentialId, String token);
-
-    // Method for generating and signing credential using remote DSS, currently not in use in DOME profile
-    Mono<String> signCredentialOnRequestedFormat(String unsignedCredential, String format, String token);
-
     Mono<Void> bindAccessTokenByPreAuthorizedCode(String processId, AuthServerNonceRequest authServerNonceRequest);
 }
