+++ src/main/java/es/in2/issuer/application/workflow/impl/CredentialSignerWorkflowImpl.java
@@ -1,18 +1,35 @@
 package es.in2.issuer.application.workflow.impl;
 
+import com.upokecenter.cbor.CBORObject;
 import es.in2.issuer.application.workflow.CredentialSignerWorkflow;
 import es.in2.issuer.application.workflow.DeferredCredentialWorkflow;
+import es.in2.issuer.domain.exception.Base45Exception;
+import es.in2.issuer.domain.model.dto.SignatureConfiguration;
+import es.in2.issuer.domain.model.dto.SignatureRequest;
 import es.in2.issuer.domain.model.dto.SignedCredentials;
+import es.in2.issuer.domain.model.dto.SignedData;
+import es.in2.issuer.domain.model.enums.SignatureType;
 import es.in2.issuer.domain.service.AccessTokenService;
 import es.in2.issuer.domain.service.CredentialProcedureService;
+import es.in2.issuer.domain.service.RemoteSignatureService;
 import es.in2.issuer.domain.util.Constants;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import nl.minvws.encoding.Base45;
+import org.apache.commons.compress.compressors.CompressorOutputStream;
+import org.apache.commons.compress.compressors.CompressorStreamFactory;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
+import reactor.core.scheduler.Schedulers;
 
+import java.io.ByteArrayOutputStream;
+import java.util.Base64;
+import java.util.Collections;
 import java.util.List;
+
+import static es.in2.issuer.domain.util.Constants.CWT_VC;
+import static es.in2.issuer.domain.util.Constants.JWT_VC;
 
 @Service
 @Slf4j
@@ -22,18 +39,19 @@ public class CredentialSignerWorkflowImpl implements CredentialSignerWorkflow {
     private final DeferredCredentialWorkflow deferredCredentialWorkflow;
     private final CredentialProcedureService credentialProcedureService;
     private final AccessTokenService accessTokenService;
-    private final VerifiableCredentialIssuanceWorkflowImpl verifiableCredentialIssuanceWorkflow;
+    private final RemoteSignatureService remoteSignatureService;
 
     @Override
-    public Mono<Void> signCredential(String authorizationHeader, String procedureId) {
+    public Mono<String> signAndUpdateCredential(String authorizationHeader, String procedureId) {
         return getCredential(authorizationHeader, procedureId)
                 .flatMap(unsignedCredential -> {
                     log.info("Start get signed credential.");
-                    return getSignedCredential(unsignedCredential, authorizationHeader);
+                    return signCredentialOnRequestedFormat(unsignedCredential, JWT_VC, authorizationHeader);
                 })
                 .flatMap(signedCredential -> {
                     log.info("Update Signed Credential");
-                    return updateSignedCredential(signedCredential);
+                    return updateSignedCredential(signedCredential)
+                            .thenReturn(signedCredential);
                 })
                 .doOnSuccess(x -> log.info("Credential Signed and updated successfully."));
     }
@@ -47,13 +65,86 @@ public class CredentialSignerWorkflowImpl implements CredentialSignerWorkflow {
                 .map(credentialDetails -> credentialDetails.credential().toString());
     }
 
-    private @NotNull Mono<String> getSignedCredential(String unsignedCredential, String token) {
-        return verifiableCredentialIssuanceWorkflow.signCredentialOnRequestedFormat(unsignedCredential, Constants.JWT_VC, token);
-    }
-
     private Mono<Void> updateSignedCredential(String signedCredential) {
         List<SignedCredentials.SignedCredential> credentials = List.of(SignedCredentials.SignedCredential.builder().credential(signedCredential).build());
         SignedCredentials signedCredentials = new SignedCredentials(credentials);
         return deferredCredentialWorkflow.updateSignedCredentials(signedCredentials);
+    }
+
+
+
+
+
+
+    @Override
+    public Mono<String> signCredentialOnRequestedFormat(String unsignedCredential, String format, String token) {
+        return Mono.defer(() -> {
+            if (format.equals(JWT_VC)) {
+                log.info(unsignedCredential);
+                log.info("Signing credential in JADES remotely ...");
+                SignatureRequest signatureRequest = new SignatureRequest(
+                        new SignatureConfiguration(SignatureType.JADES, Collections.emptyMap()),
+                        unsignedCredential
+                );
+                return remoteSignatureService.sign(signatureRequest, token)
+                        .publishOn(Schedulers.boundedElastic())
+                        .map(SignedData::data);
+            } else if (format.equals(CWT_VC)) {
+                log.info(unsignedCredential);
+                return generateCborFromJson(unsignedCredential)
+                        .flatMap(cbor -> generateCOSEBytesFromCBOR(cbor, token))
+                        .flatMap(this::compressAndConvertToBase45FromCOSE);
+            } else {
+                return Mono.error(new IllegalArgumentException("Unsupported credential format: " + format));
+            }
+        });
+    }
+
+    /**
+     * Generate CBOR payload for COSE.
+     *
+     * @param edgcJson EDGC payload as JSON string
+     * @return Mono emitting CBOR bytes
+     */
+    private Mono<byte[]> generateCborFromJson(String edgcJson) {
+        return Mono.fromCallable(() -> CBORObject.FromJSONString(edgcJson).EncodeToBytes());
+    }
+
+    /**
+     * Generate COSE bytes from CBOR bytes.
+     *
+     * @param cbor  CBOR bytes
+     * @param token Authentication token
+     * @return Mono emitting COSE bytes
+     */
+    private Mono<byte[]> generateCOSEBytesFromCBOR(byte[] cbor, String token) {
+        log.info("Signing credential in COSE format remotely ...");
+        String cborBase64 = Base64.getEncoder().encodeToString(cbor);
+        SignatureRequest signatureRequest = new SignatureRequest(
+                new SignatureConfiguration(SignatureType.COSE, Collections.emptyMap()),
+                cborBase64
+        );
+        return remoteSignatureService.sign(signatureRequest, token).map(signedData -> Base64.getDecoder().decode(signedData.data()));
+    }
+
+    /**
+     * Compress COSE bytes and convert it to Base45.
+     *
+     * @param cose COSE Bytes
+     * @return Mono emitting COSE bytes compressed and in Base45
+     */
+    private Mono<String> compressAndConvertToBase45FromCOSE(byte[] cose) {
+        return Mono.fromCallable(() -> {
+            ByteArrayOutputStream stream = new ByteArrayOutputStream();
+            try (CompressorOutputStream deflateOut = new CompressorStreamFactory()
+                    .createCompressorOutputStream(CompressorStreamFactory.DEFLATE, stream)) {
+                deflateOut.write(cose);
+            } // Automatically closed by try-with-resources
+            byte[] zip = stream.toByteArray();
+            return Base45.getEncoder().encodeToString(zip);
+        }).onErrorResume(e -> {
+            log.error("Error compressing and converting to Base45: " + e.getMessage(), e);
+            return Mono.error(new Base45Exception("Error compressing and converting to Base45"));
+        });
     }
 }
