+++ src/main/java/es/in2/issuer/domain/util/Constants.java
@@ -44,4 +44,6 @@ public class Constants {
     public static final String DID_ELSI = "did:elsi:";
     public static final List<String> CREDENTIAL_CONTEXT = List.of("https://www.w3.org/ns/credentials/v2","https://dome-marketplace.eu/2022/credentials/learcredential/v1");
     public static final List<String> VERIFIABLE_CERTIFICATION_TYPE = List.of(VERIFIABLE_CERTIFICATION, VERIFIABLE_CREDENTIAL);
+    public static final String ASYNC = "A";
+    public static final String SYNC = "S";
 }
