+++ src/test/java/es/in2/issuer/application/workflow/impl/VerifiableCredentialIssuanceServiceImplTest.java
@@ -188,44 +188,6 @@ class VerifiableCredentialIssuanceServiceImplTest {
                 .verifyComplete();
     }
 
-    @Test
-    void signCredentialOnRequestedFormat_JWT_Success() {
-        String unsignedCredential = "unsignedCredential";
-        String token = "dummyToken";
-        String signedCredential = "signedJWTData";
 
-        when(remoteSignatureService.sign(any(SignatureRequest.class), eq(token)))
-                .thenReturn(Mono.just(new SignedData(SignatureType.JADES,signedCredential)));
-
-        StepVerifier.create(verifiableCredentialIssuanceWorkflow.signCredentialOnRequestedFormat(unsignedCredential, JWT_VC, token))
-                .assertNext(signedData -> assertEquals(signedCredential, signedData))
-                .verifyComplete();
-    }
-    @Test
-    void signCredentialOnRequestedFormat_CWT_Success() {
-        String unsignedCredential = "{\"data\":\"data\"}";
-        String token = "dummyToken";
-        String signedCredential = "eyJkYXRhIjoiZGF0YSJ9";
-        String signedResult = "6BFWTLRH9.Q5$VAFLGV*M7:43S0";
-
-        when(remoteSignatureService.sign(any(SignatureRequest.class), eq(token)))
-                .thenReturn(Mono.just(new SignedData(SignatureType.COSE, signedCredential)));
-
-
-        StepVerifier.create(verifiableCredentialIssuanceWorkflow.signCredentialOnRequestedFormat(unsignedCredential, CWT_VC, token))
-                .assertNext(signedData -> assertEquals(signedResult, signedData))
-                .verifyComplete();
-    }
-
-    @Test
-    void signCredentialOnRequestedFormat_UnsupportedFormat() {
-        String unsignedCredential = "unsignedCredential";
-        String token = "dummyToken";
-        String unsupportedFormat = "unsupportedFormat";
-
-        StepVerifier.create(verifiableCredentialIssuanceWorkflow.signCredentialOnRequestedFormat(unsignedCredential, unsupportedFormat, token))
-                .expectError(IllegalArgumentException.class)
-                .verify();
-    }
 
 }
